♻️ [Refactoring]: #455 Phase 6 eol_normalize カテゴリを tests/eol_normalize.rs に分離
裸 LF / CR / CRLF の LF 正規化テスト 3 件を
literal_string/tests/eol_normalize.rs に切り出し、tests.rs ハブに
`mod eol_normalize;` 宣言を追加。
